change db

+++ .net/Models/PostInCategoriesDto.cs
@@ -7,8 +7,8 @@ namespace WebApi.Entities
 {
     public partial class PostInCategoriesDto
     {
-        public long Id { get; set; }
-        public long? PostId { get; set; }
+        public int Id { get; set; }
+        public int? PostId { get; set; }
         public int? CategoryId { get; set; }
     }
 }
